Refactor the way we fetch businesses by email
Before we searched for businesses that had *@custom-domain.com if
bob@custom-domain.com responded event though we had
info@custom-domain.com in the business.email column. However, I
forgot to think about common email addresses.

If the worker stops the email flow because it found an business with
*@gmail.com, it may or may not be fetching the appropriate businesess.

Fixed by looking to see if the email domain is common. If it is, the
query will include the full email address. If not, then it will search
for just the custom domain.