Dont display table if business has no emails
In theory this should never be the case because when a business is
created there are emails that are scheduled along with it in every case.
Added this because, I suspect in the future we will need to be able to
create a business with the option of sending an email immediately or
not.