Add redesigned Business#new form
The main reason for this redesigned form is because there needed to be a
way to create a business without delivering an email. In addtion to
this, we wanted to set it up to where we could deliver emails to
existing wholesale clients, and potential clients that turned us down at
one point in time but told us to reach back out in the future.